Make `wasm_global_set` a safe API.
Make `wasm_global_set` a safe API by allowing it to return a trap
in error cases.

Traps in the wasm spec are used to report runtime errors, and the
errors `wasm_global_set` reports are things that wasm would catch
at validation time.

Nevertheless, using a trap here is consistent with the official
JS API raising exceptions on the same conditions. And, it provides
a message explaining the reason for the error.

To support users for whom mutablility and type checks are believed
to be unacceptable overhead, add a `wasm_global_set_unsafe` function
which skips those checks, at the penalty of undefined behavior.